Use saved spawn point; autosave & chunk fixes
Load spawn point from level.dat and propagate it through ServerContext so players spawn at the world spawn instead of hardcoded coords. Add read_spawn_point, wire spawn_point into codec and make_player_join, and use it for initial PlayerPositionAndLook, SpawnPosition, and entity tracking. Change world config naming in README (world -> world_name) and switch auto_save_interval to u64, passing it into the autosave task. Improve chunk/block handling: compute section bitmask and block arrays more robustly, include light/biome data correctly, simplify block loading/save logic, and add small cleanups (early Option return, clippy allow). Also tidy imports in status packet module.

+++ coral-config/src/lib.rs
@@ -40,7 +40,7 @@ pub struct WorldConfig {
     pub allow_nether: bool,
     pub allow_end: bool,
     pub enable_auto_save: bool,
-    pub auto_save_interval: i32,
+    pub auto_save_interval: u64,
 }
 
 #[derive(Debug, Deserialize, Serialize, Clone)]
